requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
structlog>=24.1.0

tortoise-orm>=0.20.0
# Uncomment for MySQL: aiomysql>=0.2.0
# Uncomment for PostgreSQL: asyncpg>=0.29.0

redis>=5.0.0
PyJWT[crypto]>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
httpx>=0.26.0



# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
